Conta/deposito/saque

+++ src/database/migrations/20220722003945-CreateAssets.js
@@ -13,7 +13,7 @@ module.exports = {
         type: Sequelize.INTEGER,
         allowNull: false,
       },
-      value: Sequelize.INTEGER,
+      value: Sequelize.DECIMAL(10,2),
     });
   },
